Added new android library module osvrcommon and moved cross-app shareable code to this module. Also WIP work to deploy native libraries and headers from osvrcommon
