+++ OSVROpenGL/app/src/main/java/com/osvr/android/gles2sample/MainActivity.java
@@ -22,8 +22,7 @@ package com.osvr.android.gles2sample;
 import android.app.Activity;
 import android.os.Bundle;
 import android.util.Log;
-
-import com.osvr.android.utils.OSVRFileExtractor;
+import com.osvr.common.util.OSVRFileExtractor;
 
 
 public class MainActivity extends Activity {
