+++ OSVROpenGL/app/src/main/java/com/osvr/android/gles2sample/MainActivityView.java
@@ -44,16 +44,13 @@ import android.graphics.SurfaceTexture;
 import android.hardware.Camera;
 import android.opengl.GLSurfaceView;
 import android.util.Log;
-
-import com.osvr.android.jni.JNIBridge;
-
 import java.io.IOException;
-
 import javax.microedition.khronos.egl.EGL10;
 import javax.microedition.khronos.egl.EGLConfig;
 import javax.microedition.khronos.egl.EGLContext;
 import javax.microedition.khronos.egl.EGLDisplay;
 import javax.microedition.khronos.opengles.GL10;
+import com.osvr.common.jni.JNIBridge;
 
 /**
  * A simple GLSurfaceView sub-class that demonstrate how to perform
