+++ OSVROpenGL/osvrcommon/build.gradle
@@ -1,0 +1,36 @@
+apply plugin: 'com.android.model.library'
+model {
+    android {
+        compileSdkVersion 25
+        buildToolsVersion "25.0.3"
+
+        defaultConfig {
+            minSdkVersion.apiLevel = 21
+            targetSdkVersion.apiLevel = 21
+            versionCode 1
+            versionName "1.0"
+
+        }
+
+        buildTypes {
+            release {
+                minifyEnabled false
+                proguardFiles.add(file('proguard-rules.pro'))
+            }
+        }
+        sources {
+            main {
+                jni {
+                    exportedHeaders {
+                        srcDir "src/main/headers"
+                    }
+                }
+            }
+        }
+    }
+}
+
+dependencies {
+    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile 'com.android.support:appcompat-v7:25.3.1'
+}
